+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/mapping/FieldType.java
@@ -122,8 +122,7 @@ public enum FieldType implements JsonEnum {
 
 	AggregateMetricDouble("aggregate_metric_double"),
 
-	DenseVector("dense_vector"),
-
+	KnnVector("knn_vector"),
 	;
 
 	private final String jsonValue;