+++ java-client/src/main/java/org/opensearch/client/opensearch/indices/IndexSettings.java
@@ -222,6 +222,12 @@ public class IndexSettings implements JsonpSerializable {
 	@Nullable
 	private final IndexSettingsMapping mapping;
 
+	@Nullable
+	private final Boolean knn;
+
+	@Nullable
+	private final Integer knnAlgoParamEfSearch;
+
 	// ---------------------------------------------------------------------------------------------
 
 	private IndexSettings(Builder builder) {
@@ -282,6 +288,8 @@ public class IndexSettings implements JsonpSerializable {
 		this.analysis = builder.analysis;
 		this.settings = builder.settings;
 		this.mapping = builder.mapping;
+		this.knn = builder.knn;
+		this.knnAlgoParamEfSearch = builder.knnAlgoParamEfSearch;
 
 	}
 
@@ -737,6 +745,22 @@ public class IndexSettings implements JsonpSerializable {
 	}
 
 	/**
+	 * API name: {@code knn}
+	 */
+	@Nullable
+	public final Boolean knn() {
+		return this.knn;
+	}
+
+	/**
+	 * API name: {@code knn.algo_param.ef_search}
+	 */
+	@Nullable
+	public final Integer knnAlgoParamEfSearch() {
+		return this.knnAlgoParamEfSearch;
+	}
+
+	/**
 	 * Serialize this object to JSON.
 	 */
 	public void serialize(JsonGenerator generator, JsonpMapper mapper) {
@@ -1032,6 +1056,16 @@ public class IndexSettings implements JsonpSerializable {
 			this.mapping.serialize(generator, mapper);
 
 		}
+		if (this.knn != null) {
+			generator.writeKey("knn");
+			generator.write(this.knn);
+
+		}
+		if (this.knnAlgoParamEfSearch != null) {
+			generator.writeKey("knn.algo_param_ef_search");
+			generator.write(this.knnAlgoParamEfSearch);
+
+		}
 
 	}
 
@@ -1209,6 +1243,12 @@ public class IndexSettings implements JsonpSerializable {
 
 		@Nullable
 		private IndexSettingsMapping mapping;
+
+		@Nullable
+		private Boolean knn;
+
+		@Nullable
+		private Integer knnAlgoParamEfSearch;
 
 		/**
 		 * API name: {@code index}
@@ -1773,6 +1813,23 @@ public class IndexSettings implements JsonpSerializable {
 
 			return new IndexSettings(this);
 		}
+
+		/**
+		 *  API name: {@code knn}
+		 */
+		public final Builder knn(@Nullable Boolean value) {
+			this.knn = value;
+			return this;
+		}
+
+		/**
+		 * API name: {@code knn.algo_param.ef_search}
+		 */
+		public final Builder knnAlgoParamEfSearch(@Nullable Integer value) {
+			this.knnAlgoParamEfSearch = value;
+			return this;
+		}
+
 	}
 
 	// ---------------------------------------------------------------------------------------------
@@ -1874,6 +1931,9 @@ public class IndexSettings implements JsonpSerializable {
 		op.add(Builder::analysis, IndexSettingsAnalysis._DESERIALIZER, "analysis", "index.analysis");
 		op.add(Builder::settings, IndexSettings._DESERIALIZER, "settings");
 		op.add(Builder::mapping, IndexSettingsMapping._DESERIALIZER, "mapping");
+		op.add(Builder::knn, JsonpDeserializer.booleanDeserializer(), "knn", "index.knn");
+		op.add(Builder::knnAlgoParamEfSearch, JsonpDeserializer.integerDeserializer(), "knn.algo_param.ef_search",
+				"index.knn.algo_param.ef_search");
 
 	}